requirements.txt
spacy>=3.7
stanza>=1.8
streamlit>=1.32
tqdm>=4.66
pandas>=2.0
python-docx>=1.1
pdfplumber>=0.11
regex>=2024.0
torch>=2.1
